+++ dashboard/views/rfq-offers/seller-offers-search.php
@@ -123,7 +123,9 @@
                             <?php } ?>
                         </ul>
                     </div>
-                    <?php if (RfqOffers::STATUS_REJECTED == $row['offer_status'] || RfqOffers::STATUS_ACCEPTED == $row['offer_status']) { ?>
+                    <?php 
+                    $buyerAcceptance = applicationConstants::YES == $row['rlo_buyer_acceptance'];
+                    if (RfqOffers::STATUS_REJECTED == $row['offer_status'] || RfqOffers::STATUS_ACCEPTED == $row['offer_status'] || $buyerAcceptance) { ?>
                         <div class="offer-block-foot">
                             <?php if (RfqOffers::STATUS_REJECTED == $row['offer_status']) { ?>
                                 <p class="note note-rejects">
@@ -133,7 +135,7 @@
                                     </svg><?php echo Labels::getLabel('MSG_THIS_OFFER_HAS_BEEN_REJECTED_BY_BUYER'); ?>
                                 </p>
                             <?php } ?>
-                            <?php if (RfqOffers::STATUS_ACCEPTED == $row['offer_status']) { ?>
+                            <?php if (RfqOffers::STATUS_ACCEPTED == $row['offer_status'] || $buyerAcceptance) { ?>
                                 <p class="note note-accepted">
                                     <svg class="svg" width="16" height="16">
                                         <use xlink:href="<?php echo CONF_WEBROOT_URL; ?>images/retina/sprite-actions.svg#info">
@@ -220,7 +222,9 @@
                             </ul>
                         </div>
                     <?php } ?>
-                    <?php if (RfqOffers::STATUS_REJECTED == $row['counter_offer_status'] || RfqOffers::STATUS_ACCEPTED == $row['counter_offer_status'] || 1 > $row['counter_offer_id']) { ?>
+                    <?php 
+                    $sellerAcceptance = applicationConstants::YES == $row['rlo_seller_acceptance'];
+                    if (RfqOffers::STATUS_REJECTED == $row['counter_offer_status'] || RfqOffers::STATUS_ACCEPTED == $row['counter_offer_status'] || 1 > $row['counter_offer_id'] || $sellerAcceptance) { ?>
                         <div class="offer-block-foot">
                             <?php if (RfqOffers::STATUS_REJECTED == $row['counter_offer_status']) { ?>
                                 <p class="note note-rejects">
@@ -230,7 +234,7 @@
                                     </svg><?php echo Labels::getLabel('MSG_THIS_OFFER_HAS_BEEN_REJECTED_BY_YOU'); ?>
                                 </p>
                             <?php } ?>
-                            <?php if (RfqOffers::STATUS_ACCEPTED == $row['counter_offer_status']) { ?>
+                            <?php if (RfqOffers::STATUS_ACCEPTED == $row['counter_offer_status'] || $sellerAcceptance) { ?>
                                 <p class="note note-accepted">
                                     <svg class="svg" width="16" height="16">
                                         <use xlink:href="<?php echo CONF_WEBROOT_URL; ?>images/retina/sprite-actions.svg#info">