+++ CHANGELOG.md
@@ -53,6 +53,7 @@
     - Bug-084136 - Need to remove Brand parameters from Inventory Missing Info when linking brand with the catalog is not mandatory
     - Bug-084664 - Add inventory button is not working properly.
     - Bug-084668 - it shows wrong information text.
+    - Bug-084662 - Offer acceptance flow is not proper in some cases
 
 ## Hit URL:
     - admin/patch-update/updateProductRating